+++ ImTK/UI/Element/Menu/MenuView.cs
@@ -8,7 +8,7 @@ using ImTK.Core;
 
 namespace ImTK.UI
 {
-    public class MenuView : VisualElement<MenuView.Style>, IMenuElement
+    public class MenuView : VisualElement<MenuView.Style>, IMenuElement, IRenderRoot
     {
         public new class Style : VisualElement.Style
         {
@@ -34,6 +34,7 @@ namespace ImTK.UI
             }
 }
 
+        public RenderListCache RenderCache { get; } = new RenderListCache();
         public string name { get; set; }
         public int priority { get; set; }
         public bool isMenuBar { get; set; } = false;
@@ -132,7 +133,7 @@ namespace ImTK.UI
 
                 // 最後統一標記 dirty，這樣 RenderList 才會被重建
                 EventDispatcher.MarkHierarchyDirty(this);
-                this.GetWindow()?.MarkRenderListDirty();
+                this.GetRenderRoot()?.RenderCache.MarkDirty();
             }
             finally
             {