+++ ImTK/UI/IRenderRoot.cs
@@ -1,0 +1,7 @@
+namespace ImTK.UI
+{
+    public interface IRenderRoot
+    {
+        RenderListCache RenderCache { get; }
+    }
+}